style: moved tasks to separate folder, moved o3de benchmark to rai_bench package
