+++ src/rai_bench/rai_bench/main.py
@@ -1,0 +1,162 @@
+# Copyright (C) 2025 Robotec.AI
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+########### EXAMPLE USAGE ###########
+import rclpy
+import logging
+import time
+import rclpy.qos
+
+from rai_bench.benchmark_model import (
+    Benchmark,
+    Scenario,
+)
+from rai_open_set_vision.tools import GetGrabbingPointTool
+
+from rai.agents.conversational_agent import create_conversational_agent
+from rai.communication.ros2.connectors import ROS2ARIConnector
+from rai.tools.ros.manipulation import GetObjectPositionsTool, MoveToPointTool
+from rai.tools.ros2.topics import GetROS2ImageTool, GetROS2TopicsNamesAndTypesTool
+from rai.utils.model_initialization import get_llm_model
+from rai_sim.simulation_bridge import Translation
+from rai_sim.o3de.o3de_bridge import (
+    O3DEngineArmManipulationBridge,
+    O3DExROS2SimulationConfig,
+    PoseModel,
+)
+from rai_bench.o3de_test_bench.tasks import GrabCarrotTask, PlaceCubesTask
+
+from pathlib import Path
+
+
+if __name__ == "__main__":
+    rclpy.init()
+    connector = ROS2ARIConnector()
+    node = connector.node
+    node.declare_parameter("conversion_ratio", 1.0)
+
+    o3de = O3DEngineArmManipulationBridge(connector)
+
+    # define model
+    llm = get_llm_model(model_type="complex_model", streaming=True)
+
+    system_prompt = """
+    You are a robotic arm with interfaces to detect and manipulate objects.
+    Here are the coordinates information:
+    x - front to back (positive is forward)
+    y - left to right (positive is right)  
+    z - up to down (positive is up)
+    Before starting the task, make sure to grab the camera image to understand the environment.
+    """
+    # define tools
+    tools = [
+        GetObjectPositionsTool(
+            connector=connector,
+            target_frame="panda_link0",
+            source_frame="RGBDCamera5",
+            camera_topic="/color_image5",
+            depth_topic="/depth_image5",
+            camera_info_topic="/color_camera_info5",
+            get_grabbing_point_tool=GetGrabbingPointTool(connector=connector),
+        ),
+        MoveToPointTool(connector=connector, manipulator_frame="panda_link0"),
+        GetROS2ImageTool(connector=connector),
+        GetROS2TopicsNamesAndTypesTool(connector=connector),
+    ]
+    # define loggers
+    log_file = "src/rai_bench/rai_bench/benchmark.log"
+    file_handler = logging.FileHandler(log_file)
+    file_handler.setLevel(logging.DEBUG)
+
+    formatter = logging.Formatter(
+        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
+    file_handler.setFormatter(formatter)
+
+    bench_logger = logging.getLogger("Benchmark logger")
+    bench_logger.setLevel(logging.INFO)
+    bench_logger.addHandler(file_handler)
+
+    agent_logger = logging.getLogger("Agent logger")
+    agent_logger.setLevel(logging.INFO)
+    agent_logger.addHandler(file_handler)
+
+    # load different scenes
+    configs_dir = "src/rai_bench/rai_bench/o3de_test_bench/configs/"
+    connector_path = configs_dir + "o3de_config.yaml"
+    one_carrot_scene_config = O3DExROS2SimulationConfig.load_config(
+        base_config_path=Path(configs_dir + "scene1.yaml"),
+        connector_config_path=Path(connector_path),
+    )
+    multiple_carrot_scene_config = O3DExROS2SimulationConfig.load_config(
+        base_config_path=Path(configs_dir + "scene2.yaml"),
+        connector_config_path=Path(connector_path),
+    )
+    red_cubes_scene_config = O3DExROS2SimulationConfig.load_config(
+        base_config_path=Path(configs_dir + "scene3.yaml"),
+        connector_config_path=Path(connector_path),
+    )
+    multiple_cubes_scene_config = O3DExROS2SimulationConfig.load_config(
+        base_config_path=Path(configs_dir + "scene4.yaml"),
+        connector_config_path=Path(connector_path),
+    )
+    # combine different scene configs with the tasks to create various scenarios
+    scenarios = [
+        Scenario(
+            task=GrabCarrotTask(logger=bench_logger),
+            scene_config=one_carrot_scene_config,
+        ),
+        Scenario(
+            task=GrabCarrotTask(logger=bench_logger),
+            scene_config=multiple_carrot_scene_config,
+        ),
+        Scenario(
+            task=GrabCarrotTask(logger=bench_logger),
+            scene_config=red_cubes_scene_config,
+        ),
+        Scenario(
+            task=PlaceCubesTask(logger=bench_logger),
+            scene_config=red_cubes_scene_config,
+        ),
+        Scenario(
+            task=PlaceCubesTask(logger=bench_logger),
+            scene_config=multiple_cubes_scene_config,
+        ),
+    ]
+
+    # custom request to arm
+    base_arm_pose = PoseModel(translation=Translation(x=0.3, y=0.0, z=0.4))
+
+    # define benchamrk
+    benchmark = Benchmark(
+        simulation_bridge=o3de,
+        scenarios=scenarios,
+        logger=bench_logger,
+    )
+    for i, s in enumerate(scenarios):
+        agent = create_conversational_agent(
+            llm, tools, system_prompt, logger=agent_logger
+        )
+        benchmark.run_next(agent=agent)
+        o3de.move_arm(
+            pose=base_arm_pose,
+            initial_gripper_state=True,
+            final_gripper_state=False,
+            frame_id="panda_link0",
+        )  # return to case position
+        time.sleep(2)  # admire the end position for a second ;)
+
+    connector.shutdown()
+    o3de.shutdown()
+    rclpy.shutdown()
