+++ src/rai_bench/rai_bench/o3de_test_bench/configs/scene4.yaml
